% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_list.R
\name{ch_list}
\alias{ch_list}
\alias{ch_list_}
\title{Get a list}
\usage{
ch_list(id, fields = NULL, exclude_fields = NULL, key = NULL,
  parse = TRUE, ...)

ch_list_(id, fields = NULL, exclude_fields = NULL, key = NULL, ...)
}
\arguments{
\item{id}{(integer/numeric) a list id. required.}

\item{fields}{(list/vector) A comma-separated list of fields to return.
Reference parameters of sub-objects with dot notation.}

\item{exclude_fields}{(list/vector) A comma-separated list of fields to
exclude. Reference parameters of sub-objects with dot notation.}

\item{key}{A Mailchimp API key. See
\url{https://developer.mailchimp.com/documentation/mailchimp/} 
to get a key}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\description{
Get a list
}
\examples{
\dontrun{
ch_list_('208fb01591')
ch_list('208fb01591')
ch_list('208fb01591', parse = FALSE)

ch_list('208fb01591', fields = "stats")
ch_list('208fb01591', fields = "stats.open_rate")
ch_list('208fb01591', fields = "campaign_defaults")
}
}
