% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_list_growth_history.R
\name{ch_list_growth_history}
\alias{ch_list_growth_history}
\title{Get a list}
\usage{
ch_list_growth_history(id, month = NULL, fields = NULL,
  exclude_fields = NULL, count = 10, offset = 0, key = NULL,
  parse = TRUE, ...)
}
\arguments{
\item{id}{(integer/numeric) a list id. required.}

\item{month}{(character) a year/month character string of the year/month
you want. optional}

\item{fields}{(list/vector) A comma-separated list of fields to return.
Reference parameters of sub-objects with dot notation.}

\item{exclude_fields}{(list/vector) A comma-separated list of fields to
exclude. Reference parameters of sub-objects with dot notation.}

\item{count}{(integer/numeric) The number of records to return.
Default value is 10.}

\item{offset}{(integer/numeric) The number of records from a collection
to skip. Iterating over large collections with this parameter can be slow.
Default value is 0.}

\item{key}{A Mailchimp API key. See
\url{https://developer.mailchimp.com/documentation/mailchimp/} 
to get a key}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\description{
Get a list
}
\examples{
\dontrun{
ch_list_growth_history('208fb01591')$history
ch_list_growth_history('208fb01591', parse = FALSE)$history
ch_list_growth_history('208fb01591', month = "2017-06")
}
}
