% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chimpr-package.R
\docType{package}
\name{chimpr-package}
\alias{chimpr-package}
\alias{chimpr}
\title{chimpr}
\description{
Mailchimp R client
}
\section{Mailchimp API Docs}{

\url{https://mailchimp.com/developer/}
}

\author{
Scott Chamberlain \email{sckott@protonmail.com}
}
\keyword{package}
