% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{chmp_lists}
\alias{chmp_lists}
\alias{chmp_lists_}
\title{Get lists}
\usage{
chmp_lists(
  conn,
  fields = NULL,
  exclude_fields = NULL,
  count = NULL,
  offset = NULL,
  before_date_created = NULL,
  since_date_created = NULL,
  before_campaign_last_sent = NULL,
  since_campaign_last_sent = NULL,
  email = NULL,
  sort_field = NULL,
  sort_dir = NULL,
  key = NULL,
  parse = TRUE,
  ...
)

chmp_lists_(
  conn,
  fields = NULL,
  exclude_fields = NULL,
  count = NULL,
  offset = NULL,
  before_date_created = NULL,
  since_date_created = NULL,
  before_campaign_last_sent = NULL,
  since_campaign_last_sent = NULL,
  email = NULL,
  sort_field = NULL,
  sort_dir = NULL,
  key = NULL,
  ...
)
}
\arguments{
\item{conn}{a connection object. see \link{ChmpClient}}

\item{fields}{(list/vector) A comma-separated list of fields to return.
Reference parameters of sub-objects with dot notation.}

\item{exclude_fields}{(list/vector) A comma-separated list of fields to
exclude. Reference parameters of sub-objects with dot notation.}

\item{count}{(integer/numeric) The number of records to return.
Default value is 10.}

\item{offset}{(integer/numeric) The number of records from a collection
to skip. Iterating over large collections with this parameter can be slow.
Default value is 0.}

\item{before_date_created}{(character) Restrict response to lists created
before the set date.}

\item{since_date_created}{(character) Restrict results to lists created
after the set date.}

\item{before_campaign_last_sent}{(character) Restrict results to lists
created before the last campaign send date.}

\item{since_campaign_last_sent}{(character) Restrict results to lists
created after the last campaign send date.}

\item{email}{(character) Restrict results to lists that include a specific
subscriber's email address.}

\item{sort_field}{(character) Returns files sorted by the specified field.
Possible Values: date_created}

\item{sort_dir}{(character) Determines the order direction for sorted
results. Possible Values: ASC, DESC}

\item{key}{A Mailchimp API key. See
\url{https://developer.mailchimp.com/documentation/mailchimp/}
to get a key}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\description{
Get lists
}
\examples{
\dontrun{
(conn <- ChmpClient$new())
chmp_lists_(conn)
chmp_lists(conn)
chmp_lists(conn, parse = FALSE)
}
}
