% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.R
\name{chmp_ping}
\alias{chmp_ping}
\title{call the ping route}
\usage{
chmp_ping(key = NULL, ...)
}
\arguments{
\item{key}{A Mailchimp API key. See
https://developer.mailchimp.com/documentation/mailchimp/
to get a key}

\item{...}{Curl options passed to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
call the ping route
}
\examples{
\dontrun{
chmp_ping()
}
}
