% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{chmp_root}
\alias{chmp_root}
\title{call the root route}
\usage{
chmp_root(fields = NULL, exclude_fields = NULL, key = NULL,
  parse = TRUE, ...)
}
\arguments{
\item{fields}{(list/vector) A comma-separated list of fields to return.
Reference parameters of sub-objects with dot notation.}

\item{exclude_fields}{(list/vector) A comma-separated list of fields to
exclude. Reference parameters of sub-objects with dot notation.}

\item{key}{A Mailchimp API key. See
\url{https://developer.mailchimp.com/documentation/mailchimp/}
to get a key}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\description{
call the root route
}
\examples{
\dontrun{
chmp_root()
chmp_root(parse = FALSE)
}
}
