% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{exports}
\alias{exports}
\title{Exports}
\usage{
exports(conn, id, status = NULL, segment = NULL, since = NULL,
  hashed = NULL, key = NULL, ...)
}
\arguments{
\item{conn}{a connection object. see \link{ChmpClient}}

\item{key}{A Mailchimp API key. See
\url{https://developer.mailchimp.com/documentation/mailchimp/}
to get a key}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}}
}
\description{
Exports
}
\examples{
\dontrun{
(conn <- ChmpClient$new())
exports(id = "<id>")
}
}
