% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chimpr-package.R
\docType{package}
\name{chimpr-package}
\alias{chimpr-package}
\alias{chimpr}
\title{chimpr - Mailchimp R client}
\description{
chimpr - Mailchimp R client
}
\section{Mailchimp API Docs}{

\url{https://developer.mailchimp.com/documentation/mailchimp/}
}

\author{
Scott Chamberlain \email{myrmecocystus+r@gmail.com}
}
\keyword{package}
