% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaigns.R
\docType{data}
\name{ChmpCampaigns}
\alias{ChmpCampaigns}
\title{Campaigns class}
\arguments{
\item{conn}{a connection object. see \link{ChmpClient}}

\item{id}{a campgaign id. optional, but required for most methods}

\item{fields}{(list/vector) A comma-separated list of fields to return.
Reference parameters of sub-objects with dot notation.}

\item{exclude_fields}{(list/vector) A comma-separated list of fields to
exclude. Reference parameters of sub-objects with dot notation.}
}
\description{
Campaigns class
}
\details{
xxx
}
\note{
not all routes are implemented thus far
}
\examples{
\dontrun{
(conn <- ChmpClient$new())
(x <- ChmpCampaigns$new(conn))
x$all()
x$all(count = 2)
x$all(count = 2, offset = 2)

(x <- ChmpCampaigns$new(conn, id = "<id>")) 
x$info()
x$info(parse = FALSE)
x$info(fields = "report_summary")
x$content()
x$feedback()
x$send_checklist()
}
}
\keyword{datasets}
