% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ChmpClient}
\alias{ChmpClient}
\title{chimpr client}
\arguments{
\item{dc}{(character) a data center. default: us7. required.}

\item{key}{(character) API key. Can either pass in here,
or set with \code{Sys.setenv()} or pass in to each function call.
optional if already set as an environment variable}
}
\description{
chimpr client
}
\details{
\strong{Methods}
\itemize{
\item \code{ping(...)}
\itemize{
\item ... Curl options passed to \link[crul:HttpClient]{crul::HttpClient}
}
\item \code{root(fields = NULL, exclude_fields = NULL, parse = TRUE)}
\itemize{
\item fields (list/vector) A comma-separated list of fields to return.
Reference parameters of sub-objects with dot notation.
\item exclude_fields (list/vector) A comma-separated list of fields to
exclude. Reference parameters of sub-objects with dot notation.
\item parse (logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}
\item ... Curl options passed to \link[crul:HttpClient]{crul::HttpClient}
}
}
}
\examples{
\dontrun{
(x <- ChmpClient$new(dc = "us7", key = "<your key>"))
(x <- ChmpClient$new())
x$dc
x$key
x$ping()
x$root()
}
}
