% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\docType{data}
\name{ChmpList}
\alias{ChmpList}
\title{Lists class}
\arguments{
\item{conn}{a connection object. see \link{ChmpClient}}

\item{id}{a mailchimp list id. required}

\item{fields}{(list/vector) A comma-separated list of fields to return.
Reference parameters of sub-objects with dot notation.}

\item{exclude_fields}{(list/vector) A comma-separated list of fields to
exclude. Reference parameters of sub-objects with dot notation.}

\item{month}{(character) a year/month character string of the year/month
you want. optional}

\item{count}{(integer/numeric) The number of records to return.
Default value is 10.}

\item{offset}{(integer/numeric) The number of records from a collection
to skip. Iterating over large collections with this parameter can be slow.
Default value is 0.}
}
\description{
Lists class
}
\details{
xxx
}
\note{
not all routes are implemented thus far
}
\examples{
\dontrun{
(conn <- ChmpClient$new())
(x <- ChmpList$new(conn, id = "<id>"))
x$info()
x$info(parse = FALSE)
x$info(fields = "stats")
x$info(fields = "stats.open_rate")
x$info(fields = "campaign_defaults")
x$activity()
x$abuse_reports()
x$clients()
x$growth_history()
x$growth_history("2018-02")
x$interest_categories()
x$locations()
x$members()
# x$member("<id>")
# x$member_activity("<id>")
# x$member_goals("<id>")
# x$member_notes("<id>")
}
}
\keyword{datasets}
