% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{ChmpReports}
\alias{ChmpReports}
\title{Reports class}
\arguments{
\item{conn}{a connection object. see \link{ChmpClient}}

\item{id}{a report id. optional, but required for most methods}

\item{fields}{(list/vector) A comma-separated list of fields to return.
Reference parameters of sub-objects with dot notation.}

\item{exclude_fields}{(list/vector) A comma-separated list of fields to
exclude. Reference parameters of sub-objects with dot notation.}
}
\description{
Reports class
}
\details{
xxx
}
\note{
not all routes are implemented thus far
}
\examples{
\dontrun{
(conn <- ChmpClient$new())
(x <- ChmpReports$new(conn))
x$all()
x$all(count = 2)
x$all(count = 2, offset = 2)

(x <- ChmpReports$new(conn, id = "<id>"))
x$info()
x$info(parse = FALSE)
x$info(fields = "clicks")
x$abuse_reports()
x$advice()
x$open_details()
x$click_details()
# x$click_details("<id>")
# x$click_details("<id>", members = TRUE)
# x$click_details("<id>", members = TRUE, subscriber_hash = "<hash>")
x$domain_performance()
x$eepurl()
x$email_activity()
x$locations()
x$sent_to()
x$sub_reports()
x$unsubscribed()
}
}
